authapp/requirements.txt
Flask
pyjwt
cryptography

